DEV: Add last_seen_reviewable_id to the users table
We're adding this column now in preparation for a future commit(s) that will
redesign the avatar/notifications menu. The reason the column is added in a
separate commit is because the redesign changes are going to be complex with a
high risk of getting (temporarily) revereted and if they included a database
migration, they wouldn't revert cleanly/easily.

Internal ticket: t65045.